Update README.md
This is a clean to-do list web app made with HTML, CSS, and JavaScript. Users can create multiple lists, add tasks, mark them as complete, and view task dates. It features pastel and dark themes, font changes, and list deletion, all in a minimal layout for an aesthetic and clean experience.